requirements.txt
pyyaml
argparse
w1thermsensor
prometheus_client
adafruit-circuitpython-ina219
